+++ brightwind/transform/transform.py
@@ -1442,7 +1442,9 @@ def offset_timestamps(data, offset, date_from=None, date_to=None, overwrite=Fals
             return df_copy.sort_index()
 
 
-def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inplace=False):
+def apply_device_orientation_offset(
+        data, measurement_station, wdir_cols=[], inplace=False, return_orientation_offset_table=False
+        ):
     """
     Applies a device orientation offset to wind direction data from remote sensing devices
     (lidar, sodar, or floating lidar) to align measurements with north.
@@ -1472,18 +1474,22 @@ def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inp
     Overlapping periods with non-zero `device_orientation_deg` values in `vertical_profiler_properties`
     are not supported and will raise an error.
 
-    :param data:                        Timeseries data.
-    :type data:                         pd.DataFrame or pd.Series
-    :param measurement_station:         A simplified object to represent the IEA Wind Task 43 WRA Data Model.
-    :type measurement_station:          bw.MeasurementStation
-    :param wdir_cols:                   Wind direction column names to apply the offset to. If empty, all wind direction 
-                                        columns in the data are used. Default is an empty list.
-    :type wdir_cols:                    list
-    :param inplace:                     If True, modifies `data` in place. If False, returns a new DataFrame/Series
-                                        with adjusted values. Default is False.
-    :type inplace:                      bool, optional
-    :return:                            Data with wind direction adjusted by the orientation offset.
-    :rtype:                             pd.DataFrame or pd.Series
+    :param data:                                Timeseries data.
+    :type data:                                 pd.DataFrame or pd.Series
+    :param measurement_station:                 A simplified object to represent the IEA Wind Task 43 WRA Data Model.
+    :type measurement_station:                  bw.MeasurementStation
+    :param wdir_cols:                           Wind direction column names to apply the offset to. If empty, all wind 
+                                                direction columns in the data are used. Default is an empty list.
+    :type wdir_cols:                            list
+    :param inplace:                             If True, modifies `data` in place. If False, returns a new 
+                                                DataFrame/Series with adjusted values. Default is False.
+    :type inplace:                              bool, optional
+    :param return_orientation_offset_table:     If True, returns a DataFrame containing the device orientation, 
+                                                logger orientation and offset used in the analysis for each relevant 
+                                                time period.
+    :type return_orientation_offset_table:      bool, optional
+    :return:                                    Data with wind direction adjusted by the orientation offset.
+    :rtype:                                     pd.DataFrame | pd.Series | Tuple[pd.DataFrame | pd.Series, pd.DataFrame]
     
     **Example usage**
     ::
@@ -1499,6 +1505,13 @@ def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inp
     ::
         bw.apply_device_orientation_offset(data, fl1, inplace=True)
         print('Wind direction device orientation offset adjustment is completed.')
+
+    Return the orientation offset table along with adjusted data::
+    ::
+        # Adjust only specific wind direction columns:
+        data_dev_orient_adj, orientation_offset_table = bw.apply_device_orientation_offset(
+            data, fl1, wdir_cols=['Dir_40m', 'Dir_50m'], return_orientation_offset_table=True
+            )
     
     """
     
@@ -1530,6 +1543,7 @@ def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inp
 
     _check_vertical_profiler_properties_overlap(measurement_station, df)
 
+    rows = []
     # Apply the offset
     for i, wdir_prop in enumerate(wdirs_properties):
         name = wdir_prop['name']
@@ -1598,7 +1612,15 @@ def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inp
                         apply_offset_to = min(date_to_tmp, meas_station_data_model_to)
                     df[name] = _apply_dir_offset_target_orientation(
                         df[name], logger_offset, device_orientation_deg, apply_offset_from, apply_offset_to,
-                        target_orientation_name='device orientation')
+                        target_orientation_name='device orientation')            
+                    
+                    rows.append({
+                                "device_orientation": device_orientation_deg,
+                                "logger_offset": logger_offset,
+                                "offset_in_analysis": device_orientation_deg - logger_offset,
+                                "date_from": apply_offset_from,
+                                "date_to": apply_offset_to
+                            })
         else:
             wdir_not_in_dataset = True
             col_not_in_data.append(name)
@@ -1619,6 +1641,15 @@ def apply_device_orientation_offset(data, measurement_station, wdir_cols=[], inp
     if isinstance(data, pd.Series):
         df = df[df.columns[0]]
         data.update(df)
+
+    if return_orientation_offset_table:
+        device_orientation_offsets = pd.DataFrame(rows).sort_values(by=["date_from"]).groupby(
+            ["device_orientation", "logger_offset", "offset_in_analysis"]
+            ).agg({
+                "date_from": "first",
+                "date_to": lambda x: None if any(d is None for d in x) else max(x)
+                }).reset_index()
+        return df, device_orientation_offsets
     return df